+++ cpp/DBBD/DBBD/Serialize.cpp
@@ -3,9 +3,9 @@
 
 namespace DBBD {
 	void Serialize::write(Buffer& buffer, const std::string& value) {
-		size_t strSize = value.length();
+		unsigned int strSize = (unsigned int)value.length();
 		write(buffer, strSize);
-		for (size_t i = 0; i < strSize; i++) {
+		for (unsigned int i = 0; i < strSize; i++) {
 			char data = value[i];
 			buffer.putByte(data);
 		}
@@ -17,9 +17,9 @@ namespace DBBD {
 	}
 
 	void Serialize::write(Buffer& buffer, char* const& value) {
-		size_t strSize = strlen(value);
+		unsigned int strSize = (unsigned int)strlen(value);
 		write(buffer, strSize);
-		for (size_t i = 0; i < strSize; i++) {
+		for (unsigned int i = 0; i < strSize; i++) {
 			char data = value[i];
 			buffer.putByte(data);
 		}
@@ -31,7 +31,7 @@ namespace DBBD {
 
 	template<typename T1, typename T2>
 	void Serialize::writeVector(Buffer& buffer, const T1& vec) {
-		size_t vecSize = vec.size();
+		unsigned int vecSize = vec.size();
 		write(buffer, vecSize);
 		for (auto it = vec.begin(); it != vec.end(); it++) {
 			T2 data = *it;
